+++ CHANGELOG.md
@@ -10,3 +10,5 @@ similar bugs in the delay line (and all-pass & comb filters). Refactored
 the cubic interpolation and tested it seperately now.
 * Feature: Matrix::get\_connections() returns information about the connections
 to the adjacent cells.
+* Feature: Added the MatrixCellChain abstraction for easy creation of DSP
+chains on the hexagonal Matrix.